initial port to new goldmark-gemtext update
This version removes the -i and -o cli options. I'd like to implement
concurrent file/directory reading so you can just pass gemgen a list of
files and have it write the output in place. I've removed them for now
on this branch while I figure out the best way to implement the
concurrent file reading.